+++ Sources_Mod/Mod_Filter.h
@@ -6,6 +6,6 @@
 
 #define DEF_BOXCAR_WITH_MAX     100
 
-BOOL Mod_FilterBoxCar(FP64* plf_Data , INT16U uin_Len, INT16U uin_Boxcar);
+BOOL Mod_FilterBoxCar(FP32* plf_Data , INT16U uin_Len, INT16U uin_Boxcar);
 
 #endif
